Move StatsInfo to a separate component
