+++ code/client/src/components/SiderMenu.jsx
@@ -1,13 +1,10 @@
-import React, { useState, useEffect } from 'react'
+import React from 'react'
 import { useHistory, useRouteMatch } from 'react-router'
 import Layout from 'antd/es/layout'
 import Image from 'antd/es/image'
 import Row from 'antd/es/row'
 import Menu from 'antd/es/menu'
 import Typography from 'antd/es/typography'
-import Divider from 'antd/es/divider'
-import Tag from 'antd/es/tag'
-import Spin from 'antd/es/spin'
 import PlusCircleOutlined from '@ant-design/icons/PlusCircleOutlined'
 import UnorderedListOutlined from '@ant-design/icons/UnorderedListOutlined'
 import HistoryOutlined from '@ant-design/icons/HistoryOutlined'
@@ -28,11 +25,9 @@ import config from '../config'
 import Paths from '../constants/paths'
 import styled from 'styled-components'
 import { useWindowDimensions } from '../util'
-import abbr from '../abbr'
-import { useDispatch, useSelector } from 'react-redux'
-import { getColorPalette, getPrimaryBorderColor, getPrimaryTextColor } from '../theme'
-import WalletConstants from '../constants/wallet'
-import { cacheActions } from '../state/modules/cache'
+import { useSelector } from 'react-redux'
+import { getColorPalette, getPrimaryTextColor } from '../theme'
+import { StatsInfo, LineDivider } from './StatsInfo'
 const { Link } = Typography
 
 const SiderLink = styled(Link).attrs((e) => ({
@@ -49,16 +44,6 @@ const SiderLink = styled(Link).attrs((e) => ({
 const mobileMenuItemStyle = {
   padding: '0 10px',
   fontSize: 12
-}
-
-const LineDivider = ({ children }) => {
-  const theme = useSelector(state => state.global.v2ui ? state.global.theme : 'dark')
-  const color = getPrimaryBorderColor(theme)
-  return (
-    <Divider style={{ borderColor: color, opacity: 0.5, color: color, fontSize: 14 }}>
-      {children}
-    </Divider>
-  )
 }
 
 const MobileSiderMenu = ({ action, nav, ...args }) => {
@@ -79,49 +64,6 @@ const MobileSiderMenu = ({ action, nav, ...args }) => {
       <Menu.Item key='wiki' style={mobileMenuItemStyle} icon={<InfoCircleOutlined />}><SiderLink style={{ color: null }} href='https://github.com/polymorpher/one-wallet/wiki'>Wiki</SiderLink></Menu.Item>
       <Menu.Item key='tools' style={mobileMenuItemStyle} icon={<ToolOutlined />}>Tools</Menu.Item>
     </Menu>
-  )
-}
-
-const StatsInfo = () => {
-  const v2ui = useSelector(state => state.global.v2ui)
-  const statsCached = useSelector(state => state.cache.global.stats)
-  const [stats, setStats] = useState(null)
-  const dispatch = useDispatch()
-
-  useEffect(() => {
-    setStats(statsCached)
-  }, [statsCached])
-
-  useEffect(() => {
-    const { timeUpdated } = statsCached || {}
-    if (!timeUpdated || (Date.now() - timeUpdated > WalletConstants.globalStatsCacheDuration)) {
-      dispatch(cacheActions.fetchGlobalStats())
-    } else {
-      setStats(statsCached)
-    }
-  }, [])
-
-  return (
-    <>
-      <LineDivider>Global Usage</LineDivider>
-      {stats
-        ? (
-          <Row style={{ marginBottom: 16 }} justify='center'>
-            <Row style={{ marginBottom: 8 }}>
-              <Tag color='dimgray' style={{ margin: 0, width: 64, borderRadius: 0, textAlign: 'center' }}>wallets</Tag>
-              <Tag color='lightseagreen' style={{ width: 80, borderRadius: 0, textAlign: 'center' }}>{stats.count.toLocaleString()}</Tag>
-            </Row>
-            <Row>
-              <Tag color='dimgray' style={{ margin: 0, width: 64, borderRadius: 0, textAlign: 'center' }}>balance</Tag>
-              <Tag color='steelblue' style={{ width: 80, borderRadius: 0, textAlign: 'center' }}>{abbr(stats.totalAmount, 0)} ONE</Tag>
-            </Row>
-          </Row>)
-        : (
-          <Row justify='center'>
-            <Spin />
-          </Row>)}
-      {!v2ui && <LineDivider />}
-    </>
   )
 }
